+++ Products/CMFPlone/browser/sitelogo.py
@@ -1,0 +1,26 @@
+from Products.CMFPlone.interfaces import ISiteSchema
+from plone.formwidget.namedfile.converter import b64decode_file
+from plone.namedfile.browser import Download
+from plone.namedfile.file import NamedImage
+from plone.registry.interfaces import IRegistry
+from zope.component import getUtility
+
+
+class SiteLogo(Download):
+
+    def __init__(self, context, request):
+        super(SiteLogo, self).__init__(context, request)
+        self.filename = None
+        self.data = None
+
+        registry = getUtility(IRegistry)
+        settings = registry.forInterface(ISiteSchema, prefix="plone")
+        if getattr(settings, 'site_logo', False):
+            filename, data = b64decode_file(settings.site_logo)
+            data = NamedImage(data=data, filename=filename)
+            self.data = data
+            self.filename = filename
+            # self.width, self.height = self.data.getImageSize()
+
+    def _getFile(self):
+        return self.data