+++ Products/CMFPlone/interfaces/controlpanel.py
@@ -3,10 +3,10 @@ from Products.CMFPlone import PloneMessageFactory as _
 from Products.CMFPlone.utils import validate_json
 from basetool import IPloneBaseTool
 from plone.locking.interfaces import ILockSettings
+from zope import schema
 from zope.interface import Interface
 from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
-from zope import schema
 
 
 class IControlPanel(IPloneBaseTool):
@@ -865,6 +865,12 @@ class ISiteSchema(ILockSettings):
             u"This shows up in the title bar of "
             u"browsers and in syndication feeds."),
         default=u'Plone site')
+
+    site_logo = schema.ASCII(
+        title=_(u"Site Logo"),
+        description=_(u"This shows a custom Logo on your Site."),
+        required=False,
+    )
 
     exposeDCMetaTags = schema.Bool(
         title=_(u"Expose Dublin Core metadata"),